+++ tools/distrib/python/grpcio_tools/grpc_tools/command.py
@@ -27,13 +27,48 @@
 # (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 # OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
-from libc cimport stdlib
+import os
+import sys
 
-cdef extern from "grpc/tools/main.h":
-  int protoc_main(int argc, char *argv[])
+import setuptools
 
-def run_main(list args not None):
-  cdef char **argv = <char **>stdlib.malloc(len(args)*sizeof(char *))
-  for i in range(len(args)):
-    argv[i] = args[i]
-  return protoc_main(len(args), argv)
+from grpc_tools import protoc
+
+
+def build_package_protos(package_root):
+  proto_files = []
+  inclusion_root = os.path.abspath(package_root)
+  for root, _, files in os.walk(inclusion_root):
+    for filename in files:
+      if filename.endswith('.proto'):
+        proto_files.append(os.path.abspath(os.path.join(root, filename)))
+
+  for proto_file in proto_files:
+    command = [
+        'grpc.tools.protoc',
+        '--proto_path={}'.format(inclusion_root),
+        '--python_out={}'.format(inclusion_root),
+        '--grpc_python_out={}'.format(inclusion_root),
+    ] + [proto_file]
+    if protoc.main(command) != 0:
+      sys.stderr.write('warning: {} failed'.format(command))
+
+
+class BuildPackageProtos(setuptools.Command):
+  """Command to generate project *_pb2.py modules from proto files."""
+
+  description = 'build grpc protobuf modules'
+  user_options = []
+
+  def initialize_options(self):
+    pass
+
+  def finalize_options(self):
+    pass
+
+  def run(self):
+    # due to limitations of the proto generator, we require that only *one*
+    # directory is provided as an 'include' directory. We assume it's the '' key
+    # to `self.distribution.package_dir` (and get a key error if it's not
+    # there).
+    build_package_protos(self.distribution.package_dir[''])
