+++ tools/distrib/python/grpcio_tools/grpc_tools/protoc.py
@@ -1,3 +1,5 @@
+#!/usr/bin/env python
+
 # Copyright 2016, Google Inc.
 # All rights reserved.
 #
@@ -27,4 +29,21 @@
 # (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 # OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
-__import__('pkg_resources').declare_namespace(__name__)
+import pkg_resources
+import sys
+
+from grpc_tools import _protoc_compiler
+
+def main(command_arguments):
+  """Run the protocol buffer compiler with the given command-line arguments.
+
+  Args:
+    command_arguments: a list of strings representing command line arguments to
+        `protoc`.
+  """
+  command_arguments = [argument.encode() for argument in command_arguments]
+  return _protoc_compiler.run_main(command_arguments)
+
+if __name__ == '__main__':
+  proto_include = pkg_resources.resource_filename('grpc.tools', '_proto')
+  sys.exit(main(sys.argv + ['-I{}'.format(proto_include)]))
